- fixed logo and icons - added option -spendzeroconfchange to workaround tx malleability problem     as in https://github.com/bitcoin/bitcoin/pull/3651 - added seed nodes - added checkpoints - Chinese translations fixed - difficulty field in getinfo, getmininginfo, getdifficulty now return     the same as getwork: the size of the primes measured in bits. Note     that this is not linear: 1500 is not 3x harder than 1000. It's like     700 times harder. - fixed DoS (crash) vulnerability in PoW calculation

+++ src/rpcwallet.cpp
@@ -78,7 +78,7 @@ Value getinfo(const Array& params, bool fHelp)
     obj.push_back(Pair("timeoffset",    (boost::int64_t)GetTimeOffset()));
     obj.push_back(Pair("connections",   (int)vNodes.size()));
     obj.push_back(Pair("proxy",         (proxy.first.IsValid() ? proxy.first.ToStringIPPort() : string())));
-    obj.push_back(Pair("difficulty",    (int)GetDifficulty()));
+    obj.push_back(Pair("difficulty",    GetDifficulty()));
     obj.push_back(Pair("testnet",       fTestNet));
     obj.push_back(Pair("keypoololdest", (boost::int64_t)pwalletMain->GetOldestKeyPoolTime()));
     obj.push_back(Pair("keypoolsize",   pwalletMain->GetKeyPoolSize()));
